<fix> : <Article image mapping 추가 및 jpa ddl validate 수정> [지라 이슈 링크]
- article image mapping table 추가 및 ddl validation 수정
- message id 타입 int -> long 수정

+++ src/main/java/droni/backend/api/expert/entity/DroniExpert.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class DroniExpert {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long expertId;
+    private Integer expertId;
     @OneToOne
     @JoinColumn(name = "user_id", nullable = false)
     private DroniUser user;
